requirements.txt
requests
beautifulsoup4
pysocks
python-telegram-bot
python-dotenv
pybase64
